pyproject.toml
[project]
name = "learningforoptimizing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.4.0",
    "numpy>=2.1.0",
    "python-dotenv",
    "torch-geometric>=2.5.3",
    "icecream>=2.1.3",
    "sumtree>=0.2.7",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src/python"]
